+++ src/data/extension/byte_types.rs
@@ -1,8 +1,14 @@
-#![allow(dead_code)]
+//! A private module for helper functions on [`u8`].
 
-use crate::data::DNA_PROFILE_MAP;
+use crate::{
+    data::DNA_PROFILE_MAP,
+    prelude::{IsValidDNA, RecodeDNAStrat, RefineDNAStrat},
+};
 
+/// An extension trait allowing for bytes to be converted to indices with
+/// [`DNA_PROFILE_MAP`].
 pub(crate) trait ByteMappings {
+    /// Converts the byte to an index using [`DNA_PROFILE_MAP`].
     fn to_dna_index(self) -> usize;
 }
 
@@ -13,9 +19,12 @@ impl ByteMappings for u8 {
     }
 }
 
-#[allow(clippy::wrong_self_convention)]
-pub(crate) trait IsBase {
+/// Extension trait to verify whether a byte is uppercase `ACGT` or not.
+pub(crate) trait IsBase: Copy {
+    /// Returns `true` if the base is uppercase `ACGT`.
     fn is_base_acgt(self) -> bool;
+
+    /// Returns `true` if the base is not uppercase `ACGT`.
     fn is_not_base_acgt(self) -> bool;
 }
 
@@ -28,5 +37,48 @@ impl IsBase for u8 {
     #[inline]
     fn is_not_base_acgt(self) -> bool {
         !self.is_base_acgt()
+    }
+}
+
+/// A trait allowing for validation, recoding, and refinement of a single DNA
+/// base.
+///
+/// Similar functionality for entire sequences is in [`CheckNucleotides`],
+/// [`RecodeNucleotides`], and [`RetainNucleotides`]. This trait can be useful
+/// for more customized scenarios, such as recoding an array/codon via `map`.
+///
+/// [`CheckNucleotides`]: crate::data::types::nucleotides::CheckNucleotides
+/// [`RecodeNucleotides`]: crate::data::types::nucleotides::RecodeNucleotides
+/// [`RetainNucleotides`]: crate::data::types::nucleotides::RetainNucleotides
+pub trait SanitizeBase: Sized + Copy {
+    /// Checks whether a single byte is valid under the given validation
+    /// strategy
+    #[must_use]
+    fn is_valid(self, strategy: IsValidDNA) -> bool;
+
+    /// Recodes a single byte using the given strategy.
+    #[must_use]
+    fn recode_base(self, strategy: RecodeDNAStrat) -> u8;
+
+    /// Refines and recodes a single byte using the given strategy.
+    #[must_use]
+    fn refine_base(self, strategy: RefineDNAStrat) -> Option<u8>;
+}
+
+impl SanitizeBase for u8 {
+    #[inline]
+    fn is_valid(self, strategy: IsValidDNA) -> bool {
+        strategy.mapping()[self as usize]
+    }
+
+    #[inline]
+    fn recode_base(self, strategy: RecodeDNAStrat) -> u8 {
+        strategy.mapping()[self as usize]
+    }
+
+    #[inline]
+    fn refine_base(self, strategy: RefineDNAStrat) -> Option<u8> {
+        let recoded = strategy.mapping()[self as usize];
+        if recoded > 0 { Some(recoded) } else { None }
     }
 }