+++ src/data/extension/mod.rs
@@ -1,6 +1,5 @@
 /// A private module with helper functions for things that look like `[u8; N]`.
 pub(crate) mod array_types;
-/// A private module for helper functions on [`u8`].
 pub(crate) mod byte_types;
 /// A private module for working with IDs.
 pub(crate) mod id_types;