+++ src/data/mod.rs
@@ -161,6 +161,7 @@ pub use constants::mappings::{
     AA_ALL_AMBIG_PROFILE_MAP_WITH_STOP, AA_PROFILE_MAP, AA_UNAMBIG_PROFILE_MAP, ByteIndexMap, DNA_PROFILE_MAP,
     DNA_UNAMBIG_PROFILE_MAP, StdGeneticCode,
 };
+pub use extension::byte_types::SanitizeBase;
 pub use matrices::WeightMatrix;
 pub use records::{fasta, fastq, sam};
 pub use types::{amino_acids, cigar, nucleotides, phred};